Added lock() and unlock() to MenuScene
